+++ .env.deploy
@@ -4,6 +4,3 @@
 ENV_MODE=deploy
 PORT=8080
 USE_DOCKER=True
-
-BLOG_BUCKET_NAME=redline-fitness-results
-FEEDBACK_BUCKET_NAME=redline-fitness-results
