+++ .gitignore
@@ -1,13 +1,16 @@
 *.pyc
 *.pdf
-*.png
 *.log
 *.lock
 *feedback.csv
 *duration*.csv
 *df*.csv
-*.log.*
-src/blog_data/*
-src/blog_data/*.*
+src/blog_data/*/images/*.png
+src/blog_data/*/images/*.jpg
 run-scripts/gcloud-env.sh
-src/local_blog_data/*
+run-scripts/gcloud-setup.sh
+src/static/png/comp/*
+src/static/png/generic/*
+src/static/csv/generic/duration*.csv
+src/static/csv/generic/durationMensSinglesCompetitive2023.csv
+src/static/csv/generic/durationWomensSinglesOpen2023.csv
