+++ requirements.txt
@@ -5,10 +5,10 @@ Flask-Limiter==3.5.0
 gunicorn==21.2.0
 pandas==2.2.2
 Werkzeug==2.3.8
-itsdangerous==2.1.2
-click==8.1.7
+#itsdangerous==2.1.2
+#click==8.1.7
 Jinja2==3.1.3
-WTForms==3.1.2
+#WTForms==3.1.2
 python-dotenv==1.1.0
 matplotlib==3.9.0
 seaborn==0.13.2
