+++ run-scripts/push-deploy.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 # Deploy docker image to the gcloud platform 
 
-export NAME_VERSION=blog_v3.02
-export IMAGE=gcr.io/redline-fitness-results/app.py:$NAME_VERSION
+export NAME_VERSION=blog_v3.04
+export PROJECT_ID=redline-fitness-results
+export IMAGE=gcr.io/$PROJECT_ID/app.py:$NAME_VERSION
 
 #Change to the project root directory
 cd "$(dirname "$0")/.."
@@ -11,9 +12,7 @@ cd "$(dirname "$0")/.."
 docker push  $IMAGE
 
 #Deploy to Cloud Run
-#currently deploying manually via site
+#currently deploying manually via site, initially used command belwo
 #gcloud run deploy redline-results --image $IMAGE --platform managed --region asia-southeast1 --allow-unauthenticated --memory 1024M --service-account=$SERVICE_ACCOUNT
 
-#Set environment variables in the app gcloud run service itself
-#   --set-env-vars "ADMIN_PASSWORD=VALUE1" \
-#   --set-env-vars "SECRET_KEY=VALUE2" \
+
