+++ src/blog_data/2023-workout-stations-and-movements/content.json
@@ -1,0 +1,28 @@
+{
+    "headline": "2023 Workout Stations and Movements",
+    "text": "<p>This is reference post containing the 2023 Workout Stations and Movements.</p>\r\n<p>This was the OG Event, and I didnt participate. In ways was a big trial event for what it has become today.</p>\r\n<p>The 2023 details are NOT current available on the <a href=\"https://redlinefitnessgames.com/\">REDLINE Fitness Games</a> at time of writing. So I am providing her for reference purposes.</p>\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p> Thanks, </p>\r\n<p> Steve </p>",
+    "created_at": "2025-05-17T02:43:09.121828",
+    "updated_at": "2025-05-23T22:47:32.542793",
+    "published_at": "2025-05-18T21:29:48.373452",
+    "is_featured": false,
+    "slug": "2023-workout-stations-and-movements",
+    "is_published": true,
+    "view_count": 4,
+    "images": [
+        {
+            "filename": "img_20250517024308410382_0.png",
+            "caption": "",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024308736189_1.png",
+            "caption": "",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024308949760_2.png",
+            "caption": "",
+            "show_in_gallery": true
+        }
+    ]
+}
