+++ src/blog_data/2024-workout-stations-and-movements/content.json
@@ -1,0 +1,33 @@
+{
+    "headline": "2024 Workout Stations and Movements",
+    "text": "<p>This is reference post containing the 2024 Workout Stations and Movements.</p>\r\n<p>My feeling is that 2024 is where the games became and adult. This was my first time participating.&nbsp;</p>\r\n<p>The 2024 details are NOT current available on the <a href=\"https://redlinefitnessgames.com/\">REDLINE Fitness Games</a> at time of writing. So I am providing her for reference purposes.</p>\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p> Thanks, </p>\r\n<p> Steve </p>",
+    "created_at": "2025-05-17T02:45:12.479701",
+    "updated_at": "2025-05-23T22:47:27.996553",
+    "published_at": "2025-05-18T21:30:05.932812",
+    "is_featured": true,
+    "slug": "2024-workout-stations-and-movements",
+    "is_published": true,
+    "view_count": 11,
+    "images": [
+        {
+            "filename": "img_20250517024511346599_0.png",
+            "caption": "",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024511563919_1.png",
+            "caption": "",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024511847057_2.png",
+            "caption": "",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024512179323_3.png",
+            "caption": "",
+            "show_in_gallery": true
+        }
+    ]
+}
