+++ src/blog_data/2025-workout-stations-and-movements/content.json
@@ -1,0 +1,38 @@
+{
+    "headline": "2025 Workout Stations and Movements",
+    "text": "<p>This is reference post containing the 2025 Workout Stations and Movements.</p>\r\n<p>Events wises there is some name changing going on especially around the Singles, Mixed Doubles, and Mixed Relay.</p>\r\n<ul>\r\n<li><strong>Singles</strong> Level 3 ( Advanced) was previously Singles Competitive.</li>\r\n<li><strong>Singles</strong> Leve 2 (Intermediate) was previously Singles Open.</li>\r\n<li><strong>Singles</strong> Level 1 (Beginner) was previously Beginners. Singles Level 1 now has separate Men's and Women's Categorys compated to '24.</li>\r\n</ul>\r\n<ul>\r\n<li>The Double Events naming hasnt changed, but in the <strong>Mixed Doubles</strong> in '25 the Movement standards are customised. In '24 they were Mixed Standards were all the same as the Men's Standards.</li>\r\n</ul>\r\n<ul>\r\n<li>The <strong>Mixed Relay</strong> in '25 has also customised the standards differently compated to '24. Some are commen for Men &amp; Women and Some are different based on gender.</li>\r\n</ul>\r\n<p>And of course there are some new Movement Standards. In with the Shuttle Runs, Russain Twists and Squat Thrusts, and out with the Pendulum Swings, Battle Rope Whips and Agility Chamber></p>\r\n<p>Deadball burpee over target has become Deadball burpees. In '25 we need to get <em>\"the dead ball has passed on or over the marked line\" while remaining \"1m from the wall.\"&nbsp;</em></p>\r\n<p>Of course, 2025 details are still available on the <a href=\"https://redlinefitnessgames.com/\">REDLINE Fitness Games</a> at time of writing.</p>\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p>Thanks,</p>\r\n<p>Steve</p>",
+    "created_at": "2025-05-17T02:48:03.168236",
+    "updated_at": "2025-05-23T15:22:02.746152",
+    "published_at": "2025-05-18T22:10:05.678617",
+    "is_featured": true,
+    "slug": "2025-workout-stations-and-movements",
+    "is_published": true,
+    "view_count": 15,
+    "images": [
+        {
+            "filename": "img_20250517024801834078_0.png",
+            "caption": "Singles Level 1 (Beginner)",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024802136861_1.png",
+            "caption": "Singles Level 2 (Intermediate)",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024802457713_2.png",
+            "caption": "Singles Level 3 (Advanced)",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024802674229_3.png",
+            "caption": "Doubles",
+            "show_in_gallery": true
+        },
+        {
+            "filename": "img_20250517024803002123_4.png",
+            "caption": "Relay",
+            "show_in_gallery": true
+        }
+    ]
+}
