+++ src/blog_data/blog_config.json
@@ -1,0 +1,3 @@
+{
+    "max_featured_posts_on_home": 5
+}
