+++ src/blog_data/lessons-learned-from-women-s-singles-open-2024/content.json
@@ -1,0 +1,38 @@
+{
+    "headline": "Lessons Learned from Women's Singles Open 2024",
+    "text": "<p>So, let's take a look at the Women's Singles Open 2024.</p>\r\n<p>What lessons can we learn from looking at the data, and what can be gleamed to help improve in '25.&nbsp;<br /><br />All the Graphs shown here, and more, are available <a href=\"/display?eventname=WomensSinglesOpen2024\">here</a>: Just Select 'Data Visualisation' and 'Displayed on webpage'.</p>\r\n<img src=\"/blog/image/lessons-learned-from-women-s-singles-open-2024/thumb_img_20250522142828862588_0.png\" alt=\"Figure 1 (Pie Chart)\" - click to expand>\r\n<p><strong>Figure 1 (Pie Chart):</strong> This is a breakdown of the average time for each station. So you can see that the 3 longest in time are&nbsp;</p>\r\n<ol>\r\n<li>Deadball Burpee 4:42</li>\r\n<li>Sled Push Pull 4:13</li>\r\n<li>Farmer's Carry 3:57</li>\r\n</ol> \r\n<p>So from this we would think that in General these are the stations to focus on for better results.</p>\r\n<img src=\"/blog/image/lessons-learned-from-women-s-singles-open-2024/thumb_img_edit_20250523231127065798_0.png\" alt=\"Figure 2 (Bar Chart) - click to expand\" >\r\n<p><strong>Figure 2 (Bar Chart):</strong> This is a breakdown of the time spent on each station. Let me explain: If you look at the green bar for each station, it marks where top 10% to 30% of compeititors finished for that station. Some of the Green bars are small, such as for the Run; This means that the time between the 10% competitor and the 30% competitors is small. Some of the Green bars are large, such as for the Deadball burpees. You can argue that bigger the bar, the more oppurtunity there is to make up time, if you focus on it during training.</p>\r\n<p>So, let's focus on the 'middle 10% to 70% range, Green, Orange and Red bar. The 4 Stations with the longest bars for 10% to 70% are</p>\r\n<ol>\r\n<li>Deadball Burpees</li>\r\n<li>Battle Whips</li>\r\n<li>Farmer's Carry.</li>\r\n<li>Sled Push Pull</li>\r\n</ol>\r\n<img src=\"/blog/image/lessons-learned-from-women-s-singles-open-2024/thumb_img_20250522142829123711_2.png\" alt=\"Figure 3 (Correlation Bar Chart) - click to expand\" >\r\n<p><strong>Figure 3 (Correlation Bar Chart):</strong> This chart has a lot of statistics behind it, but it is looking for how Station Time is linked to the Overall Time. The Higher the number the stronger the link or Correlation.&nbsp;&nbsp;</p>\r\n<ol>\r\n<li>Farmer's Carry</li>\r\n<li>Sandbag Gauntlet</li>\r\n<li>Deadball Burpee</li>\r\n<li>Sled Push Pull</li>\r\n</ol>\r\n<p>The is a famous data science saying which goes \"Correlation is not equal to causation.\" This means that there may be a strong link between Farmer's Carry and Total Time. but it doesn't necessarily mean that  Farmer's Carry is the solution to all your problems;</p>\r\n<img src=\"/blog/image/lessons-learned-from-women-s-singles-open-2024/thumb_img_20250522142829238722_3.png\" alt=\"Figure 4 (Scatter Chart Farmer's Carry) - click to expand\" >\r\n<p><strong>Figure 4 (Scatter Chart Farmer's Carry):</strong> Each dot signifies a competitor with Overall Position on one axis and Farmer's Carry Station Time on the other. So you can see there is a pattern around a line. Note I put this line where I think it fits, but I didnt calculate it....</p>\r\n<img src=\"/blog/image/lessons-learned-from-women-s-singles-open-2024/thumb_img_20250522142829342195_4.png\" alt=\"Figure 5 (Scatter Chart  Run) - click to expand\" >\r\n<p><strong>Figure 5 (Scatter Chart Farmer's Run):</strong> Each dot signifies a competitor with Overall Position on one axis and Run Station Time on the other. You can see that there is very little pattern, so the correlation is weak.</p>\r\n<h5>Conclusion</h5>\r\n<p>There is no magic Bullet answer, but from above it would seem sensible to focus training the following events.</p>\r\n<ol>\r\n<li>Deadball Burpee</li>\r\n<li>Farmer's Carry</li>\r\n<li>Sled Push Pull</li>\r\n<li>Sandbag Gauntlet</li>\r\n</ol>\r\n<p>Battle Whips are of course gone in '25. And \"Deadball Burpee over Target\" is changing to \"Deadball Burpees\", no target anymore.....just 1 metre away, but with more reps. Farmer's Carry has increased distance from 120m to 180m. And Ski, Bike, Row all have gone from 20cal to 25cal.</p>\r\n<p>In short expect the unexpected.</p>\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p>Thanks,</p>\r\n<p>Steve</p>",
+    "created_at": "2025-05-22T14:28:29.457571",
+    "updated_at": "2025-05-23T23:35:38.351232",
+    "published_at": "2025-05-22T18:01:49.188916",
+    "is_featured": true,
+    "is_published": true,
+    "view_count": 18,
+    "slug": "lessons-learned-from-women-s-singles-open-2024",
+    "images": [
+        {
+            "filename": "img_20250522142828862588_0.png",
+            "caption": "Figure 1 (Pie Chart)",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250522142829123711_2.png",
+            "caption": "Figure 3 (Correlation Bar Chart)",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250522142829238722_3.png",
+            "caption": "Figure 4 (Scatter Chart Farmer's Carry)",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250522142829342195_4.png",
+            "caption": "Figure 5 (Scatter Chart Farmer's Run)",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_edit_20250523231127065798_0.png",
+            "caption": "Figure 2 (Bar Chart)",
+            "show_in_gallery": false
+        }
+    ]
+}
