+++ src/blog_data/race-strategy-24-men-s-single-open-winner/content.json
@@ -1,0 +1,33 @@
+{
+    "headline": "Race strategy - '24 Men's Single Open Winner",
+    "text": "<p>So what can we use all the information on this site for I hear you ask?</p>\r\n<br/>\r\nWell, let's take an example by considering the '24 Men's Single Open as it had the most entrants. We will focus on the winner, Jamie Barnett, because who doesn't want to understand a winning strategy.  I'll take it slow, so I can explain how to use the site. </p>\r\n<br/>\r\n<p>The Figures mentioned below are at the end of this page \ud83e\udd23</p>\r\n<img src=\"/blog/image/race-strategy-24-men-s-single-open-winner/thumb_img_20250516143549872045_0.png\" alt=\"Figure 1: Searching.... - click to expand\"  width=\"50%\" border=\"2\">\r\n<strong>Figure 1:</strong> \r\n<ol>\r\n<li>Click on Search Menu; </li>\r\n<li>Type Jamie Barnett into Search Bar; </li>\r\n<li>Click on Search Bar; </li>\r\n<li>Select Men's Singles Open;</li>\r\n</ol>\r\n<img src=\"/blog/image/race-strategy-24-men-s-single-open-winner/thumb_img_20250516143550008180_1.png\" alt=\"Figure 2: Selecting... - click to expand\"  width=\"50%\" border=\"2\" >\r\n\r\n<strong>Figure 2:</strong> \r\n<ol>\r\n<li>Select Displayed on this webpage; </li>\r\n<li>Click Generate. (You might need to wait a bit here as your requested data is being generated);</li>\r\n</ol>\r\n<img src=\"/blog/image/race-strategy-24-men-s-single-open-winner/thumb_img_20250516143550081901_2.png\" alt=\"Figure 3: Rank Table - click to expand\" >\r\n\r\n<strong>Figure 3:</strong> \r\n<ol>\r\n<li>You can see that our winner came in 177th place for the mule; </li>\r\n<li>You can see he was the fastest in the Sled Push Pull;</li>\r\n</ol>\r\n\r\n<img src=\"/blog/image/race-strategy-24-men-s-single-open-winner/thumb_img_20250516143550155979_3.png\" alt=\"Figure 4: Bar Chart - click to expand\">\r\n\r\n<strong>Figure 4:</strong> \r\n<ol>\r\n<li>Here you can see that our winner was in the 70th to 90th percentile for the Mule; </li>\r\n<li>Here you can see that our winner was the fastest on the Sled Push Pull;</li>\r\n</ol>\r\n\r\n<p>For the 'lazy' people at the back just click on this <a href=\"/display_vis?competitor=JAMIE%20BARNETT&year=&race_no=G1759&event=MensSinglesOpen2024\">link</a> to skip Figure 1 related steps above.</p>\r\n<p>So here you can see that there is an easy strategy to win the '24 Men's Single Open. Just be the fastest person on 6 of the 12 stations, and they you can take a rest on the 2nd last station, so that you can proceed to crush the final station!</p>\r\n\r\n<h5>Conclusion</h5>\r\n<p>I'm having a bit of fun here, but if you compare the Mule and Sled Push Pull on figure 4, you can see that it was definitely better to rank better on the sled push pull and rank lower on the mule, than vice versa, if you are able. Finding the right balance is the key.</p>\r\n\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p> Thanks, </p>\r\n<p> Steve </p>\r\n \r\n",
+    "created_at": "2025-05-16T14:35:50.304649",
+    "updated_at": "2025-05-23T22:47:10.338028",
+    "published_at": "2025-05-18T21:31:42.407347",
+    "is_featured": true,
+    "slug": "race-strategy-24-men-s-single-open-winner",
+    "is_published": true,
+    "view_count": 24,
+    "images": [
+        {
+            "filename": "img_20250516143549872045_0.png",
+            "caption": "Figure 1: Searching....",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250516143550008180_1.png",
+            "caption": "Figure 2: Selecting...",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250516143550081901_2.png",
+            "caption": "Figure 3: Rank Table",
+            "show_in_gallery": false
+        },
+        {
+            "filename": "img_20250516143550155979_3.png",
+            "caption": "Figure 4: Bar Chart",
+            "show_in_gallery": false
+        }
+    ]
+}
