+++ src/blog_data/time-to-sleep/content.json
@@ -1,0 +1,18 @@
+{
+    "headline": "Time to sleep",
+    "text": "Time to Sleep\r\n\r\n<p>(if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs) Some text before the no align image.  Some text before the no align image.  Some text before the no align image.  Some text before the no align image. Some text before the no align image. \r\n    <img src=\"/blog/image/time-to-sleep/thumb_img_edit_20250522231340421809_0.jpg\" alt=\"This is the Caption\" width=\"40%\" border=\"2\">\r\n(if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs)Some text after the no align image.</p>\r\n\r\n\r\n<p>So, let's take a look at the Women's Singles Open 2024.</p>\r\n<p>What lessons can we learn from looking at the data, and what can be gleamed to help improve in '25.&nbsp;<br /><br />All the Graphs shown here, and more, are available <a href=\"/display?eventname=WomensSinglesOpen2024\">here</a>: Just Select 'Data Visualisation' and 'Displayed on webpage'.</p>\r\n<p><strong>Figure 1 (Pie Chart):</strong> This is a breakdown of the average time for each station. So you can see that the 3 longest in time are&nbsp;</p>\r\n<ol>\r\n<li>Deadball Burpee 4:42</li>\r\n<li>Sled Push Pull 4:13</li>\r\n<li>Farmer's Carry 3:57</li>\r\n</ol>\r\n<p>So from this we would think that in General these are the stations to focus on for better results.</p>\r\n<p><strong>Figure 2 (Bar Chart):</strong> This is a breakdown of the time spent on each station. Let me explain: If you look at the green bar for each station, it marks where top 10% to 30% of compeititors finished for that station. Some of the Green bars are small, such as for the Run; This means that the time between the 10% competitor and the 30% competitors is small. Some of the Green bars are large, such as for the Deadball burpees. You can argue that bigger the bar, the easier it might be to make up time if you focus on it during training.</p>\r\n<p>So, let's focus on the 'middle 10% to 70% range, Green, Orange and Red bar. The 3 Stations with the longest bars for 10% to 70% are</p>\r\n<ol>\r\n<li>Deadball Burpees</li>\r\n<li>Battle Whips</li>\r\n<li>Farmer's Carry.</li>\r\n</ol>\r\n<p><strong>Figure 3 (Correlation Bar Chart):</strong> This chart has a lot of statistics behind it, but it is looking for how Station Time is linked to the Overall Time. The Higher the number the stronger the link or Correlation.&nbsp;&nbsp;</p>\r\n<ol>\r\n<li>Farmer's Carry</li>\r\n<li>Sandbag Gauntlet</li>\r\n<li>Deadball Burpee</li>\r\n<li>Sled Push Pull</li>\r\n</ol>\r\n<p>The is a famous data science saying which goes \"Correlation is not equal to causation.\" This means that there may be a strong link between Farmer's Carry and Total Time. but it doesn't necessarily mean that  Farmer's Carry is the solution to all your problems;</p>\r\n<p><strong>Figure 4 (Scatter Chart Farmer's Carry):</strong> Each dot signifies a competitor with Overall Position on one axis and Farmer's Carry Station Time on the other. So you can see there is a pattern around a line. Note I put this line where I think it fits, but I didnt calculate it....</p>\r\n<p><strong>Figure 5 (Scatter Chart Farmer's Run):</strong> Each dot signifies a competitor with Overall Position on one axis and Run Station Time on the other. You can see that there is very little pattern, so the correlation is weak.</p>\r\n<h5>Conclusion</h5>\r\n<p>There is no magic Bullet answer, but from above it would seem sensible to focus training the following events.</p>\r\n<ol>\r\n<li>Deadball Burpee&nbsp;</li>\r\n<li>Farmer's Carry</li>\r\n<li>Sandbag Gauntlet</li>\r\n<li>Sled Push Pull</li>\r\n</ol>\r\n<p>Battle Whips are of course gone in '25. And \"Deadball Burpee over Target\" is changing to \"Deadball Burpees\", no target anymore.....just 1 metre away, but with more reps. Farmer's Carry has increased distance from 120m to 180m. And Ski, Bike, Row all have gone from 20cal to 25cal.</p>\r\n<p>In short expect the unexpected.</p>\r\n<p>If you have any questions, requests or need any explanations, please stick it in the <a href=\"/feedback\">feedback form</a>. and I'll do my best to address.</p>\r\n<p>Thanks,</p>\r\n<p>Steve</p>\r\n\r\n\r\n\r\n\r\n\r\n\r\n\r\nform_data (if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs) Ensure your Python backend logic for edit_blog_post correctly reads current_show_in_gallery_{idx} and new_show_in_gallery_{i} from request.form. Ensure your Python backend logic for new_blog_post correctly reads show_in_gallery_{i} from request.form. This template should now provide the necessary checkboxes for your \"show in gallery\" feature. .</p>\r\n\r\n\r\n\r\n<p>form_data (if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs) Ensure your Python backend logic for edit_blog_post correctly reads current_show_in_gallery_{idx} and new_show_in_gallery_{i} from request.form. Ensure your Python backend logic for new_blog_post correctly reads show_in_gallery_{i} from request.form. This template should now provide the necessary checkboxes for your \"show in gallery\" feature. \r\n\r\n<p>(if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs) Some text before the image. \r\n<a><img src=\"/blog/image/time-to-sleep/thumb_img_edit_20250522231340421809_0.jpg\" alt=\"This will become the caption!\"></a>\r\n(if you use it for repopulating on validation error \u2013 the image part of this would be complex to repopulate fully for checkboxes and file inputs)Some text after the image.</p>\r\n",
+    "images": [
+        {
+            "filename": "img_20250522230246936926_0.jpg",
+            "caption": "Image 1",
+            "show_in_gallery": true
+        }
+    ],
+    "created_at": "2025-05-22T23:02:47.095037",
+    "updated_at": "2025-05-23T22:57:00.393082",
+    "published_at": "2025-05-22T23:02:47.095037",
+    "is_featured": false,
+    "is_published": false,
+    "view_count": 27,
+    "slug": "time-to-sleep"
+}
